Refactor file structure to increase modular structure
