+++ src/Containers/MovieContainer/MovieContainer.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import './MovieContainer.css';
-import MovieCard from '../Components/MovieCard';
+import './MovieContainer/MovieContainer.css';
+import MovieCard from '../Components/MovieCard/MovieCard';
 
 const MovieContainer = ({movies}) => {
